WIP: Start GTK4 support

+++ lgi/override/Gdk.lua
@@ -19,9 +19,12 @@ local ti = ffi.types
 local Gdk = lgi.Gdk
 local cairo = lgi.cairo
 
--- Take over internal GDK synchronization lock.
-core.registerlock(core.gi.Gdk.resolve.gdk_threads_set_lock_functions)
-Gdk.threads_init()
+-- Take over internal GDK synchronization lock in Gdk3.
+-- This API does not exist in Gdk4.
+if core.gi.Gdk.resolve.gdk_threads_set_lock_functions then
+   core.registerlock(core.gi.Gdk.resolve.gdk_threads_set_lock_functions)
+   Gdk.threads_init()
+end
 
 -- Gdk.Rectangle does not exist at all in older GOI, because it is
 -- aliased to cairo.RectangleInt.  Make sure that we have it exists,
@@ -33,27 +36,29 @@ if not Gdk.Rectangle then
    Gdk.Rectangle._method.union = Gdk.rectangle_union
 end
 
--- Declare GdkAtoms which are #define'd in Gdk sources and not
+-- Declare GdkAtoms which are #define'd in Gdk3 sources and not
 -- introspected in gir.
-local _ = Gdk.KEY_0
-for name, val in pairs {
-   SELECTION_PRIMARY = 1,
-   SELECTION_SECONDARY = 2,
-   SELECTION_CLIPBOARD = 69,
-   TARGET_BITMAP = 5,
-   TARGET_COLORMAP = 7,
-   TARGET_DRAWABLE = 17,
-   TARGET_PIXMAP = 20,
-   TARGET_STRING = 31,
-   SELECTION_TYPE_ATOM = 4,
-   SELECTION_TYPE_BITMAP = 5,
-   SELECTION_TYPE_COLORMAP = 7,
-   SELECTION_TYPE_DRAWABLE = 17,
-   SELECTION_TYPE_INTEGER = 19,
-   SELECTION_TYPE_PIXMAP = 20,
-   SELECTION_TYPE_WINDOW = 33,
-   SELECTION_TYPE_STRING = 31,
-} do Gdk._constant[name] = Gdk.Atom(val) end
+if Gdk.Atom then
+   local _ = Gdk.KEY_0
+   for name, val in pairs {
+      SELECTION_PRIMARY = 1,
+      SELECTION_SECONDARY = 2,
+      SELECTION_CLIPBOARD = 69,
+      TARGET_BITMAP = 5,
+      TARGET_COLORMAP = 7,
+      TARGET_DRAWABLE = 17,
+      TARGET_PIXMAP = 20,
+      TARGET_STRING = 31,
+      SELECTION_TYPE_ATOM = 4,
+      SELECTION_TYPE_BITMAP = 5,
+      SELECTION_TYPE_COLORMAP = 7,
+      SELECTION_TYPE_DRAWABLE = 17,
+      SELECTION_TYPE_INTEGER = 19,
+      SELECTION_TYPE_PIXMAP = 20,
+      SELECTION_TYPE_WINDOW = 33,
+      SELECTION_TYPE_STRING = 31,
+   } do Gdk._constant[name] = Gdk.Atom(val) end
+end
 
 -- Easier-to-use Gdk.RGBA.parse() override.
 if Gdk.RGBA then
@@ -72,18 +77,22 @@ end
 
 -- Gdk.Window.destroy() actually consumes 'self'.  Prepare workaround
 -- with override doing ref on input arg first.
-local destroy = Gdk.Window.destroy
-local ref = core.callable.new {
-   addr = core.gi.GObject.resolve.g_object_ref,
-   ret = ti.ptr, ti.ptr
-}
-function Gdk.Window:destroy()
-   ref(self._native)
-   destroy(self)
+if Gdk.Window then
+   local destroy = Gdk.Window.destroy
+   local ref = core.callable.new {
+      addr = core.gi.GObject.resolve.g_object_ref,
+      ret = ti.ptr, ti.ptr
+   }
+   function Gdk.Window:destroy()
+      ref(self._native)
+      destroy(self)
+   end
 end
 
 -- Better integrate Gdk cairo helpers.
-Gdk.Window.cairo_create = Gdk.cairo_create
+if Gdk.Window then
+   Gdk.Window.cairo_create = Gdk.cairo_create
+end
 cairo.Region.create_from_surface = Gdk.cairo_region_create_from_surface
 
 local cairo_set_source_rgba = cairo.Context.set_source_rgba
